+++ DataFormats/TrackerRecHit2D/interface/TrackingRecHitLessFromGlobalPosition.h
@@ -4,7 +4,7 @@
 #include "DataFormats/TrajectorySeed/interface/PropagationDirection.h"
 #include "Geometry/CommonDetUnit/interface/GeomDet.h"
 #include <functional>
-#include "Geometry/TrackerGeometryBuilder/interface/GeomDetLess.h"
+#include "Geometry/CommonDetUnit/interface/GeomDetLess.h"
 #include "Geometry/CommonDetUnit/interface/TrackingGeometry.h"
 #include "DataFormats/TrackingRecHit/interface/TrackingRecHit.h"
 /** Defines order of layers in the Tracker as seen by straight tracks
