+++ DataFormats/TrackerRecHit2D/src/ProjectedSiStripRecHit2D.cc
@@ -1,5 +1,5 @@
 #include "DataFormats/TrackerRecHit2D/interface/ProjectedSiStripRecHit2D.h"
-#include "Geometry/TrackerGeometryBuilder/interface/GluedGeomDet.h"
+#include "Geometry/CommonDetUnit/interface/GluedGeomDet.h"
 
 // #include<iostream>
