+++ Geometry/CommonDetUnit/src/GeomDetLess.cc
@@ -1,4 +1,4 @@
-#include "Geometry/TrackerGeometryBuilder/interface/GeomDetLess.h"
+#include "Geometry/CommonDetUnit/interface/GeomDetLess.h"
 #include "DataFormats/SiStripDetId/interface/StripSubdetector.h"
 #include "DataFormats/SiPixelDetId/interface/PixelSubdetector.h"
 #include "FWCore/Utilities/interface/Exception.h"
