+++ RecoTracker/MeasurementDet/plugins/TkGluedMeasurementDet.cc
@@ -1,7 +1,7 @@
 #include "TkGluedMeasurementDet.h"
 #include "TrackingTools/MeasurementDet/interface/MeasurementDetException.h"
 #include "RecoTracker/TransientTrackingRecHit/interface/TSiStripMatchedRecHit.h"
-#include "Geometry/TrackerGeometryBuilder/interface/GluedGeomDet.h"
+#include "Geometry/CommonDetUnit/interface/GluedGeomDet.h"
 #include "RecoLocalTracker/SiStripRecHitConverter/interface/SiStripRecHitMatcher.h"
 #include "NonPropagatingDetMeasurements.h"
 #include "TrackingTools/PatternTools/interface/TrajectoryMeasurement.h"
